dpi and alpha correction

+++ src/shaders/shader.vert
@@ -7,7 +7,7 @@ layout (set = 0, binding = 0) uniform Globals {
 
 layout(location = 0) in vec2 a_pos;
 layout(location = 1) in vec2 a_tc;
-layout(location = 2) in uvec4 a_color;
+layout(location = 2) in vec4 a_color;
 
 out gl_PerVertex { vec4 gl_Position; };
 layout(location = 0) out vec2 v_tc;
